feat: resize memorized, webview hook refactor, add jsdoc, update readme
