+++ src/client/types.ts
@@ -1,9 +1,21 @@
 export type IData = {
+  /** webview url */
   url: string;
+  /** webview url hostname */
   hostname?: string;
+  /** webview title */
   title?: string;
+  /** webview icon (the most clear) */
   icon?: string;
+  /** webview icon (alternative) */
   altIcon?: string;
+  /** webview favicon */
   favicon?: string;
+  /** webview manifest link */
   manifest?: string;
+  /** trayWindow memorized bounds */
+  bounds?: {
+    width: number;
+    height: number;
+  };
 };
